largest number

+++ 0x03-debugging/2-largest_number.c
@@ -8,7 +8,7 @@
  * Return: largest number
  */
 
-int largest_number(int a, int b, int c)
+double largest_number(double a, double b, double c)
 {
 	int largest;
